finished

+++ user-service/src/main/java/com/tarikkamat/userservice/dto/ReviewDTO.java
@@ -7,6 +7,6 @@ public record ReviewDTO(
         String userId,
         String restaurantId,
         String comment,
-        Integer rating
+        Double rating
 ) {
 }
